modify 'MusicPlayer'
- change method: seekTo(long) -> seekTo(int)

+++ player/src/main/java/snow/player/media/MediaMusicPlayer.java
@@ -80,7 +80,7 @@ public class MediaMusicPlayer extends MusicPlayer {
     }
 
     @Override
-    public void seekTo(long pos) {
+    public void seekTo(int pos) {
         mMediaPlayer.seekTo((int) pos);
     }
